\name{abricateR}
\alias{abricateR}
\title{abricateR: a package for the processing of abricate, pMLST, pointfinder data.}
\usage{
abricateR(abricate_in,
                 output,
                 output_directory,
                 identity,
                 length,
                 writecsv,
                 pointfinder_data,
                 pMLST_data)
}
\description{
A function for processing abricate output and combining it with pointfinder and pMLST data. Also generates data on ColV carriage based on the Liu criteria.
}
